dont chunl a wobbly if no lock when unlocking

+++ src/system_registry/system_registry.rb
@@ -462,7 +462,7 @@ class SystemRegistry < Registry
   end
   
   def unlock_tree
-    File.delete(@@RegistryLock)
+    File.delete(@@RegistryLock) if File.exist?(@@RegistryLock)
   end
   
     # @sets the service_tree and load mod time
